🔨🤖 move shared props handling into AbstractTokenTextWrap
MarkdownTextWrap and TextWrapGroup duplicated the same defaultOptions,
constructor and props/options getters. Make AbstractTokenTextWrap
generic over its props type and lift the shared block into it, so
subclasses only provide their own fields plus text/tokens.
